Bugfix: 修正了使用二进制字符串作为红黑树的 key 时，红黑树无法进行正确处理的 BUG

+++ src/common/wbt_file.h
@@ -37,7 +37,7 @@ typedef struct wbt_file_s {
     /* 文件大小 */
     size_t size;
     /* 已发送数据偏移量 */
-    size_t offset;
+    off_t offset;
     /* 最近使用时间 */
     time_t last_use_mtime;
     /* 最近修改时间 */
